#47 refactor: fixture 분리
