+++ be/issue/src/test/java/codesquad/issueTracker/comment/controller/CommentControllerTest.java
@@ -7,18 +7,15 @@ import static org.springframework.test.web.servlet.request.MockMvcRequestBuilder
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
 
 import annotation.ControllerTest;
-import codesquad.issueTracker.comment.domain.Comment;
 import codesquad.issueTracker.comment.dto.CommentRequestDto;
 import codesquad.issueTracker.comment.dto.CommentResponseDto;
+import codesquad.issueTracker.comment.fixture.CommentTestFixture;
 import codesquad.issueTracker.comment.service.CommentService;
-import codesquad.issueTracker.comment.vo.CommentUserVo;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import java.time.LocalDateTime;
 import java.util.List;
-import java.util.stream.Collectors;
-import java.util.stream.IntStream;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -30,9 +27,12 @@ import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
 
 
 @ControllerTest(CommentController.class)
-class CommentControllerTest {
+class CommentControllerTest extends CommentTestFixture {
 
     private final Log log = LogFactory.getLog(CommentControllerTest.class);
+
+    private List<CommentResponseDto> commentResponseDtosFixture;
+    private CommentRequestDto commentRequestDtoFixture;
 
     @Autowired
     private MockMvc mockMvc;
@@ -43,11 +43,17 @@ class CommentControllerTest {
     @MockBean
     private CommentService commentService;
 
+    @BeforeEach
+    public void setUp() {
+        commentResponseDtosFixture = dummyCommentResponseDto();
+        commentRequestDtoFixture = dummyCommentRequestDto();
+    }
+
     @Test
     @DisplayName("모든 댓글 목록들을 반환한다.")
     public void getComments() throws Exception {
         //given
-        given(commentService.getComments(any())).willReturn(dummyCommentResponseDto());
+        given(commentService.getComments(any())).willReturn(commentResponseDtosFixture);
 
         //when
         ResultActions resultActions = mockMvc.perform(get("/api/issues/{issueId}/comments", 1L));
@@ -56,50 +62,11 @@ class CommentControllerTest {
         resultActions
                 .andExpect(status().isOk())
                 .andExpect(jsonPath("$.status").value(SUCCESS.getStatus().getReasonPhrase()))
-                .andExpect(jsonPath("$.message[0].id").value(1))
-                .andExpect(jsonPath("$.message[0].createdAt").value(dummyLocalDateTime().toString()))
-                .andExpect(jsonPath("$.message[0].content").value("comment content1"))
-                .andExpect(jsonPath("$.message[0].writer.name").value(makeCommentUser().getName()))
-                .andExpect(jsonPath("$.message[0].writer.profileImg").value(makeCommentUser().getProfileImg()));
-
-    }
-
-    private List<CommentResponseDto> dummyCommentResponseDto() {
-        return IntStream.range(1, 4)
-                .mapToObj(this::makeCommentResponses)
-                .collect(Collectors.toList());
-    }
-
-    private CommentResponseDto makeCommentResponses(int num) {
-        Comment comment = makeComment(num);
-        CommentUserVo commentUserVo = makeCommentUser();
-
-        return CommentResponseDto.builder()
-                .id((long) num)
-                .content(comment.getContent())
-                .writer(commentUserVo)
-                .createdAt(comment.getCreatedAt())
-                .build();
-    }
-
-    private Comment makeComment(int num) {
-        return Comment.builder()
-                .issueId((long) num)
-                .userId(1L)
-                .content("comment content" + num)
-                .createdAt(dummyLocalDateTime())
-                .build();
-    }
-
-    private LocalDateTime dummyLocalDateTime() {
-        return LocalDateTime.of(2023, 8, 12, 7, 23, 10);
-    }
-
-    private CommentUserVo makeCommentUser() {
-        return CommentUserVo.builder()
-                .name("sio")
-                .profileImg("https://upload.wikimedia.org/wikipedia/commons/1/17/Enhydra_lutris_face.jpg")
-                .build();
+                .andExpect(jsonPath("$.message[0].id").value(commentResponseDtosFixture.get(0).getId()))
+                .andExpect(jsonPath("$.message[0].createdAt").value(commentResponseDtosFixture.get(0).getCreatedAt().toString()))
+                .andExpect(jsonPath("$.message[0].content").value(commentResponseDtosFixture.get(0).getContent()))
+                .andExpect(jsonPath("$.message[0].writer.name").value(commentResponseDtosFixture.get(0).getWriter().getName()))
+                .andExpect(jsonPath("$.message[0].writer.profileImg").value(commentResponseDtosFixture.get(0).getWriter().getProfileImg()));
     }
 
     @Test
@@ -110,17 +77,14 @@ class CommentControllerTest {
 
         //when
         ResultActions resultActions = mockMvc.perform(post("/api/issues/{issueId}/comments", 1L)
-                .content(objectMapper.writeValueAsString(dummyCommentRequestDto()))
-                .contentType(MediaType.APPLICATION_JSON));
+                .content(objectMapper.writeValueAsString(commentRequestDtoFixture))
+                .contentType(MediaType.APPLICATION_JSON)
+                .requestAttr("userId", 1L));
 
         //then
         resultActions.andExpect(status().isOk())
                 .andExpect(jsonPath("$.status").value(SUCCESS.getStatus().getReasonPhrase()))
                 .andExpect(jsonPath("$.message").value(SUCCESS.getMessage()));
-    }
-
-    private CommentRequestDto dummyCommentRequestDto() {
-        return new CommentRequestDto("post comment test1");
     }
 
     @Test
@@ -131,7 +95,7 @@ class CommentControllerTest {
 
         //when
         ResultActions resultActions = mockMvc.perform(patch("/api/issues/comments/{commentId}", 1L)
-                .content(objectMapper.writeValueAsString(dummyCommentRequestDto()))
+                .content(objectMapper.writeValueAsString(commentRequestDtoFixture))
                 .contentType(MediaType.APPLICATION_JSON));
 
         //then
